+++ CHANGELOG.md
@@ -15,6 +15,7 @@ details.
 ### Changed
 - Adjusted strict mode setting for ES5 configuration.
 - checkrules.js: show pre-release rule changes from github master branch.
+- Allow template literals in `quotes` rule.
 
 ### Fixed
 - `eslint` 2.13.0 and later require relative extend to include file extension.